Refactor encryption and decryption logic for messages.
Replaces `EncryptionKey` with `LinkKey` for clarity and updates associated methods accordingly. Refactors encryption key length handling to use `uint16` instead of `int64`. Removes redundant configuration and improves code readability with streamlined APIs and error handling.

+++ config.go
@@ -6,8 +6,8 @@ import (
 )
 
 type MessageConfig struct {
-	MinEncryptionKeyLength int64
-	MaxEncryptionKeyLength int64
+	MinEncryptionKeyLength uint16
+	MaxEncryptionKeyLength uint16
 	MaxTextLength          int64
 }
